+++ src/index.ts
@@ -28,6 +28,7 @@ export type {
 } from './components/Select';
 export * from './components/Surface';
 export * from './components/LoadingBackdrop';
+export * from './components/Snackbar';
 export * from './components/Switch';
 export * from './components/Tabs';
 export * from './components/TextField';